expand package names
