+++ src/main/java/ru/dickpickgalactic/andrew/sudoku/Cell.java
@@ -1,4 +1,4 @@
-package ru.dickpickgalactic.sudoku;
+package ru.dickpickgalactic.andrew.sudoku;
 
 import java.util.Arrays;
 import java.util.TreeSet;
@@ -7,9 +7,6 @@ public class Cell {
 
   public Integer value = null;
   public TreeSet<Integer> suggested = new TreeSet<>(Arrays.asList(1, 2, 3, 4, 5, 6, 7, 8, 9));
-
-  public Cell() {
-  }
 
   public Cell(char value) {
     if (value != '.') {
