+++ src/main/java/ru/dickpickgalactic/andrew/sudoku/Main.java
@@ -1,4 +1,4 @@
-package ru.dickpickgalactic.sudoku;
+package ru.dickpickgalactic.andrew.sudoku;
 
 //TIP To <b>Run</b> code, press <shortcut actionId="Run"/> or
 // click the <icon src="AllIcons.Actions.Execute"/> icon in the gutter.
